This diff creates code related to EmbeddingBag in ATen. It also allows sparse gradients.

This is the command I used to create the timings for both 
```
NUMEXPR_NUM_THREADS=8 MKL_NUM_THREADS=8 OMP_NUM_THREADS=8 taskset -c 0-7 python benchmarks/embeddingbag.py
```

Master
```
""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
runs: 10000     number of indices: 2000 maximum number of bags: 200     maximum bag size: 30

====================================================================================================
dimension:      10000   x       100
----------------------------------------------------------------------------------------------------
cpu dense       cpu sparse      cuda dense      cuda sparse
  13.175s          0.000s          7.638s          0.000s

====================================================================================================
dimension:      10000   x       1000
----------------------------------------------------------------------------------------------------
cpu dense       cpu sparse      cuda dense      cuda sparse
  66.805s          0.000s          7.442s          0.000s

====================================================================================================
dimension:      100000  x       100
----------------------------------------------------------------------------------------------------
cpu dense       cpu sparse      cuda dense      cuda sparse
  59.101s          0.000s          9.520s          0.000s

====================================================================================================
dimension:      100000  x       1000
----------------------------------------------------------------------------------------------------
cpu dense       cpu sparse      cuda dense      cuda sparse
 495.547s          0.000s         30.891s          0.000s
```

emb_tmp
```
""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
runs: 10000     number of indices: 2000 maximum number of bags: 200     maximum bag size: 30

====================================================================================================
dimension:      10000   x       100
----------------------------------------------------------------------------------------------------
cpu dense       cpu sparse      cuda dense      cuda sparse
  15.195s         17.045s          6.571s          6.513s

====================================================================================================
dimension:      10000   x       1000
----------------------------------------------------------------------------------------------------
cpu dense       cpu sparse      cuda dense      cuda sparse
  64.924s         21.673s          8.399s          6.361s

====================================================================================================
dimension:      100000  x       100
----------------------------------------------------------------------------------------------------
cpu dense       cpu sparse      cuda dense      cuda sparse
  57.969s         14.924s          8.211s          6.711s

====================================================================================================
dimension:      100000  x       1000
----------------------------------------------------------------------------------------------------
cpu dense       cpu sparse      cuda dense      cuda sparse
 496.196s         22.128s         30.167s          6.683s
```

[This](https://pastebin.com/0K6aNjqS) is the script I used to benchmark.
EDIT: Added CUDA synchronize

The performance between both branches matches, but this new code has sparse gradients and lives within ATen.

The code was built with
```
python setup.py build_deps develop
```

The timings of the first 1000 runs are discarded and the benchmark is executed 10000 times.