@colesbury 2018-01-25
RE: the build errors: s/_VariableBase/_VariableFunctions