@SsnL 2018-02-07
chiming in to say that the relevant issue is #4441 , which should be closed once this PR is approved and merged.